django/requirements/base.txt
pytz
Pillow
django
django-environ
django-anymail[mailgun]
django-s3-storage
sentry-sdk
django-extensions
celery
pytest
pytest-django
pytest-xdist
coverage
whitenoise
gunicorn
psycopg2
factory-boy
ipython
argon2-cffi
djangorestframework
django-cors-headers
drf_spectacular
djoser
tablib[xlsx]
